pyproject.toml
[tool.poetry]
name = "atmacup-coffee"
version = "0.1.0"
description = ""
authors = ["azupero_adolph <azupero.adolph@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
numpy = "^1.19.5"
pandas = "^1.2.0"
matplotlib = "^3.3.3"
seaborn = "^0.11.1"
scipy = "^1.6.0"
plotly = "^4.14.3"
pygam = "^0.8.0"
scikit-learn = "^0.24.0"
jupyterlab = "^3.0.3"
ipywidgets = "^7.6.3"
tqdm = "^4.56.0"
catboost = "^0.24.4"
mlflow = "^1.13.1"
hydra-core = "^1.0.5"
category-encoders = "^2.2.2"
matplotlib-venn = "^0.11.6"
optuna = "^2.4.0"
xfeat = "^0.1.1"
mypy = "^0.790"

[tool.poetry.dev-dependencies]
pytest = "^6.2.1"
flake8 = "^3.8.4"
black = "^20.8b1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
